server/requirements.txt
# Python dependencies for analytics scripts (from your local .venv311)
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
Faker==37.8.0
fonttools==4.60.0
holidays==0.81
importlib_resources==6.5.2
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prophet==1.1.7
pymongo==4.15.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
stanio==0.5.1
tqdm==4.67.1
tzdata==2.25.2